Split up apolloHelpers into multiple files
