+++ services/app-web/src/components/SubmissionSummarySection/RateDetailsSummarySection/RateDetailsSummarySection.stories.tsx
@@ -4,7 +4,7 @@ import {
     RateDetailsSummarySectionProps,
     RateDetailsSummarySection,
 } from './RateDetailsSummarySection'
-import { mockContractAndRatesDraft } from '../../../testHelpers/apolloHelpers'
+import { mockContractAndRatesDraft } from '../../../testHelpers/apolloMocks'
 
 export default {
     title: 'Components/SubmissionSummary/RateDetailsSummarySection',
