+++ services/app-web/src/components/SubmissionSummarySection/SubmissionTypeSummarySection/SubmissionTypeSummarySection.stories.tsx
@@ -4,7 +4,7 @@ import {
     SubmissionTypeSummarySectionProps,
     SubmissionTypeSummarySection,
 } from './SubmissionTypeSummarySection'
-import { mockContractAndRatesDraft } from '../../../testHelpers/apolloHelpers'
+import { mockContractAndRatesDraft } from '../../../testHelpers/apolloMocks'
 
 export default {
     title: 'Components/SubmissionSummary/SubmissionTypeSummarySection',
