+++ services/app-web/src/pages/StateDashboard/StateDashboard.test.tsx
@@ -7,7 +7,7 @@ import {
     mockDraftHealthPlanPackage,
     mockSubmittedHealthPlanPackage,
     mockUnlockedHealthPlanPackage,
-} from '../../testHelpers/apolloHelpers'
+} from '../../testHelpers/apolloMocks'
 import {
     ldUseClientSpy,
     renderWithProviders,
