+++ services/app-web/src/pages/StateSubmission/RateDetails/RateDetails.test.tsx
@@ -16,7 +16,7 @@ import {
     mockUnlockedHealthPlanPackage,
     mockSubmittedHealthPlanPackage,
     mockMNState,
-} from '../../../testHelpers/apolloHelpers'
+} from '../../../testHelpers/apolloMocks'
 
 import {
     renderWithProviders,
