+++ services/app-web/src/pages/SubmissionSideNav/SubmissionSideNav.test.tsx
@@ -15,7 +15,7 @@ import {
     mockSubmittedHealthPlanPackage,
     mockUnlockedHealthPlanPackage,
     mockValidCMSUser,
-} from '../../testHelpers/apolloHelpers'
+} from '../../testHelpers/apolloMocks'
 import { ldUseClientSpy } from '../../testHelpers'
 
 describe('SubmissionSideNav', () => {
