+++ services/app-web/src/testHelpers/apolloMocks/stateMock.ts
@@ -1,0 +1,32 @@
+import { State } from '../../gen/gqlClient'
+
+function mockMNState(): State {
+    return {
+        name: 'Minnesota',
+        programs: [
+            {
+                id: 'abbdf9b0-c49e-4c4c-bb6f-040cb7b51cce',
+                fullName: 'Special Needs Basic Care',
+                name: 'SNBC',
+            },
+            {
+                id: 'd95394e5-44d1-45df-8151-1cc1ee66f100',
+                fullName: 'Prepaid Medical Assistance Program',
+                name: 'PMAP',
+            },
+            {
+                id: 'ea16a6c0-5fc6-4df8-adac-c627e76660ab',
+                fullName: 'Minnesota Senior Care Plus ',
+                name: 'MSC+',
+            },
+            {
+                id: '3fd36500-bf2c-47bc-80e8-e7aa417184c5',
+                fullName: 'Minnesota Senior Health Options',
+                name: 'MSHO',
+            },
+        ],
+        code: 'MN',
+    }
+}
+
+export { mockMNState }
