+++ services/app-web/src/testHelpers/apolloMocks/userGQLMock.ts
@@ -1,0 +1,61 @@
+import { MockedResponse } from '@apollo/client/testing'
+import { FetchCurrentUserDocument, User as UserType } from '../../gen/gqlClient'
+
+import { mockMNState } from './stateMock'
+function mockValidUser(): UserType {
+    return {
+        __typename: 'StateUser' as const,
+        id: 'foo-id',
+        state: mockMNState(),
+        role: 'STATE_USER',
+        givenName: 'bob',
+        familyName: 'ddmas',
+        email: 'bob@dmas.mn.gov',
+    }
+}
+
+function mockValidCMSUser(): UserType {
+    return {
+        __typename: 'CMSUser' as const,
+        id: 'bar-id',
+        role: 'CMS_USER',
+        givenName: 'bob',
+        familyName: 'ddmas',
+        email: 'bob@dmas.mn.gov',
+        stateAssignments: [],
+    }
+}
+
+type fetchCurrentUserMockProps = {
+    user?: UserType | Partial<UserType>
+    statusCode: 200 | 403 | 500
+}
+const fetchCurrentUserMock = ({
+    user = mockValidUser(),
+    statusCode,
+}: // eslint-disable-next-line @typescript-eslint/no-explicit-any
+fetchCurrentUserMockProps): MockedResponse<Record<string, any>> => {
+    switch (statusCode) {
+        case 200:
+            return {
+                request: { query: FetchCurrentUserDocument },
+                result: {
+                    data: {
+                        fetchCurrentUser: user,
+                    },
+                },
+            }
+        case 403:
+            return {
+                request: { query: FetchCurrentUserDocument },
+                error: new Error('You are not logged in'),
+            }
+        default:
+            return {
+                request: { query: FetchCurrentUserDocument },
+                error: new Error('A network error occurred'),
+            }
+    }
+}
+
+export { fetchCurrentUserMock, mockValidCMSUser, mockValidUser }
